Start to handle exceptions more correctly...
Don't rely on builtin-exception handling matching interpreter exception
handling. This shortcut works best when interpreter is same as the same
as the bytecode version, but gets worse and worse as the two get farther
appart in changes. Because Python's exception handling changs.

The exception objects can be different and the error messages can be
different.